resolve issue of expired membership

+++ includes/classes/class-discord-api.php
@@ -66,10 +66,6 @@ class PMPro_Discord_API {
 	 * @return int $curr_level_id
 	 */
 	public function get_current_level_id( $user_id ) {
-		if( !is_user_logged_in() ) {
-			wp_send_json_error( 'Unauthorized user', 404 );
-			exit();
-		}
 		if ( is_user_logged_in() && function_exists( 'pmpro_hasMembershipLevel' ) && pmpro_hasMembershipLevel() ) {
 			global $current_user;
 			$membership_level = pmpro_getMembershipLevelForUser( $user_id );
@@ -299,10 +295,6 @@ class PMPro_Discord_API {
 	 * @return object API response
 	 */
 	public function delete_member_from_guild( $user_id ) {
-		if( !is_user_logged_in() ) {
-			wp_send_json_error( 'Unauthorized user', 404 );
-			exit();
-		}
 		$guild_id = get_option( 'ets_discord_guild_id' );
 		$discord_bot_token = get_option( 'ets_discord_bot_token' );
 		$ets_discord_user_id = get_user_meta( $user_id , 'ets_discord_user_id', true );
@@ -335,10 +327,6 @@ class PMPro_Discord_API {
 	 * @return object API response
 	 */
 	public function change_discord_role_api( $user_id, $role_id ) {
-		if( !is_user_logged_in() ) {
-			wp_send_json_error( 'Unauthorized user', 404 );
-			exit();
-		}
 		$access_token = get_user_meta( $user_id, "ets_discord_access_token", true );
 		$previous_role = get_user_meta( $user_id, "ets_discord_role_id", true );
 		$guild_id = get_option( 'ets_discord_guild_id' );
@@ -377,10 +365,6 @@ class PMPro_Discord_API {
 	 * @return object API response 
 	 */
 	public function delete_discord_role( $user_id ) {
-		if( !is_user_logged_in() ) {
-			wp_send_json_error( 'Unauthorized user', 404 );
-			exit();
-		}
 		$access_token = get_user_meta( $user_id, "ets_discord_access_token", true );
 		$guild_id = get_option( 'ets_discord_guild_id' );
 		$ets_discord_user_id = get_user_meta( $user_id, 'ets_discord_user_id', true );
@@ -411,10 +395,6 @@ class PMPro_Discord_API {
 	 * @return object API response
 	 */
 	public function change_discord_role_from_pmpro( $level_id, $user_id, $cancel_level ) {
-		if( !is_user_logged_in() ) {
-			wp_send_json_error( 'Unauthorized user', 404 );
-			exit();
-		}
 		$ets_discord_user_id = get_user_meta( $user_id, 'ets_discord_user_id',true );
 		if ( $ets_discord_user_id ) {
 			$role_delete = $this->delete_discord_role( $user_id );
@@ -467,17 +447,15 @@ class PMPro_Discord_API {
 	 * @return None
 	 */
 	public function pmpro_expiry_membership( $user_id, $level_id ) {
-		if( !is_user_logged_in() ) {
-			wp_send_json_error( 'Unauthorized user', 404 );
-			exit();
-		}	
 		$ets_discord_role_mapping = json_decode(get_option( 'ets_discord_role_mapping' ), true );
 		$role_id = '';
 		$role_id = get_option('ets_discord_default_role_id');
 		$allow_none_member = get_option( 'ets_allow_none_member' );
+		$ets_discord_role_mapping = json_decode(get_option( 'ets_discord_role_mapping' ), true );
+		$curr_level_id = $this->get_current_level_id( $user_id );
 		if ($allow_none_member == 'no') {
 			$this->delete_member_from_guild( $user_id );
-		} else if ($allow_none_member == 'yes') {
+		} else if ($allow_none_member == 'yes' && !empty($role_id) && array_key_exists('level_id_'.$curr_level_id, $ets_discord_role_mapping)) {
 			$this->delete_discord_role( $user_id );
 			$this->change_discord_role_api( $user_id, $role_id );
 		}
